最近检查了下代码，BUG什么的倒是没找到，变量类型和变量命名以及函数命名有点乱，决定用cstdint和驼峰命名整个替换一下

+++ Engine/Inc/Gear/Camera/OrbitCamera.h
@@ -40,9 +40,9 @@ private:
 
 	const DirectX::XMVECTOR up;
 
-	int moveEventID;
+	uint64_t moveEventID;
 
-	int scrollEventID;
+	uint64_t scrollEventID;
 
 };
